method reorg
