+++ hapi-fhir-jpaserver-searchparam/src/main/java/ca/uhn/fhir/jpa/cache/ResourceChangeListenerCacheRefresherImpl.java
@@ -8,6 +8,7 @@ import ca.uhn.fhir.jpa.searchparam.retry.Retrier;
 import com.google.common.annotations.VisibleForTesting;
 import org.apache.commons.lang3.time.DateUtils;
 import org.hl7.fhir.instance.model.api.IBaseResource;
+import org.hl7.fhir.instance.model.api.IIdType;
 import org.quartz.JobExecutionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,7 +16,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
+import java.util.ArrayList;
 import java.util.Iterator;
+import java.util.List;
 
 @Service
 public class ResourceChangeListenerCacheRefresherImpl implements IResourceChangeListenerCacheRefresher {
@@ -93,8 +96,61 @@ public class ResourceChangeListenerCacheRefresherImpl implements IResourceChange
 		}
 		SearchParameterMap searchParamMap = theEntry.getSearchParameterMap();
 		ResourceVersionMap newResourceVersionMap = myResourceVersionSvc.getVersionMap(theEntry.getResourceName(), searchParamMap);
-		retval = retval.plus(myResourceChangeListenerRegistry.notifyListener(theEntry, newResourceVersionMap));
+		retval = retval.plus(notifyListener(theEntry, newResourceVersionMap));
 
 		return retval;
+	}
+
+	/**
+	 * Notify a listener with all matching resources if it hasn't been initialized yet, otherwise only notify it if
+	 * any resources have changed
+	 * @param theListenerEntry
+	 * @param theNewResourceVersionMap the measured new resources
+	 * @return the list of created, updated and deleted ids
+	 */
+	// FIXME KHS move notification stuff out to its own service
+	ResourceChangeResult notifyListener(RegisteredResourceChangeListener theListenerEntry, ResourceVersionMap theNewResourceVersionMap) {
+		ResourceChangeResult retval;
+		IResourceChangeListener resourceChangeListener = theListenerEntry.getResourceChangeListener();
+		if (theListenerEntry.isInitialized()) {
+			retval = compareLastVersionMapToNewVersionMapAndNotifyListenerOfChanges(resourceChangeListener, theListenerEntry.getResourceVersionCache(), theNewResourceVersionMap);
+		} else {
+			theListenerEntry.getResourceVersionCache().initialize(theNewResourceVersionMap);
+			resourceChangeListener.handleInit(theNewResourceVersionMap.getSourceIds());
+			retval = ResourceChangeResult.fromCreated(theNewResourceVersionMap.size());
+			theListenerEntry.setInitialized(true);
+		}
+		return retval;
+	}
+
+	private ResourceChangeResult compareLastVersionMapToNewVersionMapAndNotifyListenerOfChanges(IResourceChangeListener theListener, ResourceVersionCache theOldResourceVersionCache, ResourceVersionMap theNewResourceVersionMap) {
+		// If the new ResourceVersionMap does not have the old key - delete it
+		List<IIdType> deletedIds = new ArrayList<>();
+		theOldResourceVersionCache.keySet()
+			.forEach(id -> {
+				if (!theNewResourceVersionMap.containsKey(id)) {
+					deletedIds.add(id);
+				}
+			});
+		deletedIds.forEach(theOldResourceVersionCache::removeResourceId);
+
+		List<IIdType> createdIds = new ArrayList<>();
+		List<IIdType> updatedIds = new ArrayList<>();
+
+		for (IIdType id : theNewResourceVersionMap.keySet()) {
+			String previousValue = theOldResourceVersionCache.put(id, theNewResourceVersionMap.get(id));
+			IIdType newId = id.withVersion(theNewResourceVersionMap.get(id));
+			if (previousValue == null) {
+				createdIds.add(newId);
+			} else if (!theNewResourceVersionMap.get(id).equals(previousValue)) {
+				updatedIds.add(newId);
+			}
+		}
+
+		IResourceChangeEvent resourceChangeEvent = ResourceChangeEvent.fromCreatedUpdatedDeletedResourceIds(createdIds, updatedIds, deletedIds);
+		if (!resourceChangeEvent.isEmpty()) {
+			theListener.handleChange(resourceChangeEvent);
+		}
+		return ResourceChangeResult.fromResourceChangeEvent(resourceChangeEvent);
 	}
 }
