+++ hapi-fhir-jpaserver-searchparam/src/main/java/ca/uhn/fhir/jpa/cache/ResourceChangeListenerRegistryImpl.java
@@ -7,16 +7,13 @@ import ca.uhn.fhir.jpa.searchparam.matcher.InMemoryMatchResult;
 import ca.uhn.fhir.jpa.searchparam.matcher.InMemoryResourceMatcher;
 import com.google.common.annotations.VisibleForTesting;
 import org.hl7.fhir.instance.model.api.IBaseResource;
-import org.hl7.fhir.instance.model.api.IIdType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Nonnull;
-import java.util.ArrayList;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
@@ -59,7 +56,7 @@ public class ResourceChangeListenerRegistryImpl implements IResourceChangeListen
 
 	@Override
 	public void unregisterResourceResourceChangeListener(IResourceChangeListener theResourceChangeListener) {
-		remove(theResourceChangeListener);
+		myListenerEntries.removeIf(l -> l.getResourceChangeListener().equals(theResourceChangeListener));
 	}
 
 	private RegisteredResourceChangeListener add(String theResourceName, IResourceChangeListener theResourceChangeListener, SearchParameterMap theMap, long theRemoteRefreshIntervalMs) {
@@ -78,65 +75,6 @@ public class ResourceChangeListenerRegistryImpl implements IResourceChangeListen
 	@Nonnull
 	public Iterator<RegisteredResourceChangeListener> iterator() {
 		return myListenerEntries.iterator();
-	}
-
-	/**
-	 * Notify a listener with all matching resources if it hasn't been initialized yet, otherwise only notify it if
-	 * any resources have changed
-	 * @param theListenerEntry
-	 * @param theNewResourceVersionMap the measured new resources
-	 * @return the list of created, updated and deleted ids
-	 */
-	// FIXME KHS move notification stuff out to its own service
-	@Override
-	public ResourceChangeResult notifyListener(RegisteredResourceChangeListener theListenerEntry, ResourceVersionMap theNewResourceVersionMap) {
-		ResourceChangeResult retval;
-		IResourceChangeListener resourceChangeListener = theListenerEntry.getResourceChangeListener();
-		if (theListenerEntry.isInitialized()) {
-			retval = compareLastVersionMapToNewVersionMapAndNotifyListenerOfChanges(resourceChangeListener, theListenerEntry.getResourceVersionCache(), theNewResourceVersionMap);
-		} else {
-			theListenerEntry.getResourceVersionCache().initialize(theNewResourceVersionMap);
-			resourceChangeListener.handleInit(theNewResourceVersionMap.getSourceIds());
-			retval = ResourceChangeResult.fromCreated(theNewResourceVersionMap.size());
-			theListenerEntry.setInitialized(true);
-		}
-		return retval;
-	}
-
-	private ResourceChangeResult compareLastVersionMapToNewVersionMapAndNotifyListenerOfChanges(IResourceChangeListener theListener, ResourceVersionCache theOldResourceVersionCache, ResourceVersionMap theNewResourceVersionMap) {
-		// If the new ResourceVersionMap does not have the old key - delete it
-		List<IIdType> deletedIds = new ArrayList<>();
-		theOldResourceVersionCache.keySet()
-			.forEach(id -> {
-				if (!theNewResourceVersionMap.containsKey(id)) {
-					deletedIds.add(id);
-				}
-			});
-		deletedIds.forEach(theOldResourceVersionCache::removeResourceId);
-
-		List<IIdType> createdIds = new ArrayList<>();
-		List<IIdType> updatedIds = new ArrayList<>();
-
-		for (IIdType id : theNewResourceVersionMap.keySet()) {
-			String previousValue = theOldResourceVersionCache.put(id, theNewResourceVersionMap.get(id));
-			IIdType newId = id.withVersion(theNewResourceVersionMap.get(id));
-			if (previousValue == null) {
-				createdIds.add(newId);
-			} else if (!theNewResourceVersionMap.get(id).equals(previousValue)) {
-				updatedIds.add(newId);
-			}
-		}
-
-		IResourceChangeEvent resourceChangeEvent = ResourceChangeEvent.fromCreatedUpdatedDeletedResourceIds(createdIds, updatedIds, deletedIds);
-		if (!resourceChangeEvent.isEmpty()) {
-			theListener.handleChange(resourceChangeEvent);
-		}
-		return ResourceChangeResult.fromResourceChangeEvent(resourceChangeEvent);
-	}
-
-	// FIXME KHS inline
-	private void remove(IResourceChangeListener theResourceChangeListener) {
-		myListenerEntries.removeIf(l -> l.getResourceChangeListener().equals(theResourceChangeListener));
 	}
 
 	public int size() {
