+++ hapi-fhir-jpaserver-searchparam/src/test/java/ca/uhn/fhir/jpa/cache/ResourceChangeListenerRegistryImplTest.java
@@ -21,7 +21,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -31,14 +30,9 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(SpringExtension.class)
@@ -63,10 +57,8 @@ class ResourceChangeListenerRegistryImplTest {
 	@MockBean
 	private SearchParamMatcher mySearchParamMatcher;
 
-	private IResourceChangeListener myTestListener = mock(IResourceChangeListener.class);
-	private SearchParameterMap myMap = SearchParameterMap.newSynchronous();
-	private Set<RegisteredResourceChangeListener> myEntries;
-	private RegisteredResourceChangeListener myEntry;
+	private final IResourceChangeListener myTestListener = mock(IResourceChangeListener.class);
+	private static final SearchParameterMap ourMap = SearchParameterMap.newSynchronous();
 
 	@Configuration
 	@Import(RegisteredResourceListenerFactoryConfig.class)
@@ -84,16 +76,16 @@ class ResourceChangeListenerRegistryImplTest {
 
 	@BeforeEach
 	public void before() {
-		myEntries = new HashSet<>();
-		myEntry = myRegisteredResourceListenerFactory.create(PATIENT_RESOURCE_NAME, myMap, myTestListener, TEST_REFRESH_INTERVAL_MS);
-		myEntries.add(myEntry);
+		Set<RegisteredResourceChangeListener> entries = new HashSet<>();
+		RegisteredResourceChangeListener entry = myRegisteredResourceListenerFactory.create(PATIENT_RESOURCE_NAME, ourMap, myTestListener, TEST_REFRESH_INTERVAL_MS);
+		entries.add(entry);
 		when(myInMemoryResourceMatcher.checkIfInMemorySupported(any(), any())).thenReturn(InMemoryMatchResult.successfulMatch());
 	}
 
 	@Test
 	public void addingListenerForNonResourceFails() {
 		try {
-			myResourceChangeListenerRegistry.registerResourceResourceChangeListener("Foo", myMap, myTestListener, TEST_REFRESH_INTERVAL_MS);
+			myResourceChangeListenerRegistry.registerResourceResourceChangeListener("Foo", ourMap, myTestListener, TEST_REFRESH_INTERVAL_MS);
 			fail();
 		} catch (DataFormatException e) {
 			assertEquals("Unknown resource name \"Foo\" (this name is not known in FHIR version \"R4\")", e.getMessage());
@@ -104,7 +96,7 @@ class ResourceChangeListenerRegistryImplTest {
 	public void addingNonInMemorySearchParamFails() {
 		try {
 			mockInMemorySupported(InMemoryMatchResult.unsupportedFromReason("TEST REASON"));
-			myResourceChangeListenerRegistry.registerResourceResourceChangeListener(PATIENT_RESOURCE_NAME, myMap, myTestListener, TEST_REFRESH_INTERVAL_MS);
+			myResourceChangeListenerRegistry.registerResourceResourceChangeListener(PATIENT_RESOURCE_NAME, ourMap, myTestListener, TEST_REFRESH_INTERVAL_MS);
 			fail();
 		} catch (IllegalArgumentException e) {
 			assertEquals("SearchParameterMap SearchParameterMap[] cannot be evaluated in-memory: TEST REASON.  Only search parameter maps that can be evaluated in-memory may be registered.", e.getMessage());
@@ -112,7 +104,7 @@ class ResourceChangeListenerRegistryImplTest {
 	}
 
 	private void mockInMemorySupported(InMemoryMatchResult theTheInMemoryMatchResult) {
-		when(myInMemoryResourceMatcher.checkIfInMemorySupported(myMap, ourFhirContext.getResourceDefinition(PATIENT_RESOURCE_NAME))).thenReturn(theTheInMemoryMatchResult);
+		when(myInMemoryResourceMatcher.checkIfInMemorySupported(ourMap, ourFhirContext.getResourceDefinition(PATIENT_RESOURCE_NAME))).thenReturn(theTheInMemoryMatchResult);
 	}
 
 	@AfterEach
@@ -124,15 +116,15 @@ class ResourceChangeListenerRegistryImplTest {
 	@Test
 	public void registerUnregister() {
 		IResourceChangeListener listener1 = mock(IResourceChangeListener.class);
-		myResourceChangeListenerRegistry.registerResourceResourceChangeListener(PATIENT_RESOURCE_NAME, myMap, listener1, TEST_REFRESH_INTERVAL_MS);
-		myResourceChangeListenerRegistry.registerResourceResourceChangeListener(OBSERVATION_RESOURCE_NAME, myMap, listener1, TEST_REFRESH_INTERVAL_MS);
+		myResourceChangeListenerRegistry.registerResourceResourceChangeListener(PATIENT_RESOURCE_NAME, ourMap, listener1, TEST_REFRESH_INTERVAL_MS);
+		myResourceChangeListenerRegistry.registerResourceResourceChangeListener(OBSERVATION_RESOURCE_NAME, ourMap, listener1, TEST_REFRESH_INTERVAL_MS);
 
 		when(mySearchParamMatcher.match(any(), any())).thenReturn(InMemoryMatchResult.successfulMatch());
 
 		assertEquals(2, myResourceChangeListenerRegistry.size());
 
 		IResourceChangeListener listener2 = mock(IResourceChangeListener.class);
-		myResourceChangeListenerRegistry.registerResourceResourceChangeListener(PATIENT_RESOURCE_NAME, myMap, listener2, TEST_REFRESH_INTERVAL_MS);
+		myResourceChangeListenerRegistry.registerResourceResourceChangeListener(PATIENT_RESOURCE_NAME, ourMap, listener2, TEST_REFRESH_INTERVAL_MS);
 		assertEquals(3, myResourceChangeListenerRegistry.size());
 
 		List<RegisteredResourceChangeListener> entries = Lists.newArrayList(myResourceChangeListenerRegistry.iterator());
@@ -145,7 +137,7 @@ class ResourceChangeListenerRegistryImplTest {
 		assertThat(resourceNames, contains(PATIENT_RESOURCE_NAME, OBSERVATION_RESOURCE_NAME, PATIENT_RESOURCE_NAME));
 
 		RegisteredResourceChangeListener firstEntry = entries.iterator().next();
-		assertEquals(myMap, firstEntry.getSearchParameterMap());
+		assertEquals(ourMap, firstEntry.getSearchParameterMap());
 
 		myResourceChangeListenerRegistry.unregisterResourceResourceChangeListener(listener1);
 		assertEquals(1, myResourceChangeListenerRegistry.size());
@@ -154,28 +146,5 @@ class ResourceChangeListenerRegistryImplTest {
 		assertEquals(listener2, entry.getResourceChangeListener());
 		myResourceChangeListenerRegistry.unregisterResourceResourceChangeListener(listener2);
 		assertEquals(0, myResourceChangeListenerRegistry.size());
-	}
-
-	@Test
-	public void testNotifyListenersEmptyEmptyNotInitialized() {
-		IResourceChangeListener listener = mock(IResourceChangeListener.class);
-		RegisteredResourceChangeListener entry = new RegisteredResourceChangeListener(PATIENT_RESOURCE_NAME, listener, myMap, TEST_REFRESH_INTERVAL_MS);
-		ResourceVersionMap newResourceVersionMap = ResourceVersionMap.fromResourceTableEntities(Collections.emptyList());
-		assertFalse(entry.isInitialized());
-		myResourceChangeListenerRegistry.notifyListener(entry, newResourceVersionMap);
-		assertTrue(entry.isInitialized());
-		verify(listener, times(1)).handleInit(any());
-	}
-
-	@Test
-	public void testNotifyListenersEmptyEmptyInitialized() {
-		IResourceChangeListener listener = mock(IResourceChangeListener.class);
-		RegisteredResourceChangeListener entry = new RegisteredResourceChangeListener(PATIENT_RESOURCE_NAME, listener, myMap, TEST_REFRESH_INTERVAL_MS);
-		ResourceVersionMap newResourceVersionMap = ResourceVersionMap.fromResourceTableEntities(Collections.emptyList());
-		entry.setInitialized(true);
-		assertTrue(entry.isInitialized());
-		myResourceChangeListenerRegistry.notifyListener(entry, newResourceVersionMap);
-		assertTrue(entry.isInitialized());
-		verifyNoInteractions(listener);
 	}
 }
